Make registry an interface to allow different registry types
